% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucket_performance.R
\name{bucket_performance}
\alias{bucket_performance}
\title{Creates an xts of the mean forward returns per factor bucket.}
\usage{
bucket_performance(
  factor_buckets,
  forward_returns,
  average_func = c("mean", "median")
)
}
\arguments{
\item{factor_buckets}{An xts object containing the buckets for each asset.}

\item{forward_returns}{An xts object of na padded forward returns for each asset.}

\item{average_func}{A string of the  unction name used to average the forward returns.}
}
\description{
Creates an xts of the mean forward returns per factor bucket.
}
