% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_finances_data.R
\docType{data}
\name{dummy_finances}
\alias{dummy_finances}
\title{Small dataset containing fake finance data of banks in xts format. Used for testing and exploration.}
\format{
An xts object spanning from 2015-01-01 to 2023-01-01 by quarter with 21 columns.
}
\source{
https://finance.yahoo.com
}
\usage{
dummy_finances
}
\description{
Small dataset containing fake finance data of banks in xts format. Used for testing and exploration.
}
\examples{
data(dummy_finances)
}
\keyword{datasets}
