% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_bucket.R
\name{factor_bucket}
\alias{factor_bucket}
\title{Places factor data into discrete groups.}
\usage{
factor_bucket(
  factor_data,
  method = c("balanced", "proximal", "fixed"),
  width,
  intervals
)
}
\arguments{
\item{factor_data}{An xts object}

\item{method}{balanced allows for buckets of relatively equal content, most likely with some remainder if width is not a multiple of the length #'  of a row. proximal allows for bucketing based on the proximity to other values. fixed is the method that must be selected to use predefined  #'  intervals for bucketing}

\item{width}{The number of buckets. If less than one, treated as a percentage of the non NA values in any row.}

\item{intervals}{A vector of numbers representing the intervals for each bucket.}
}
\description{
Places factor data into discrete groups.
}
\note{
This function is useful for bounding ranks into set intervals, or scoring bounded measures like a correlation coefficient.
}
