% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_quantile.R
\name{factor_quantile}
\alias{factor_quantile}
\title{Buckets factors into quantiles}
\usage{
factor_quantile(
  factor_data,
  method = c("traditional", "weighted"),
  probs = seq(0, 1, 0.25),
  weights
)
}
\arguments{
\item{factor_data}{An xts object or a list of xts objects}

\item{method}{Which quantizing method to choose. Weighted uses weighted z scores to find the quantiles.}

\item{probs}{A list or vector of the probabilities of each quantile.}

\item{weights}{A vector containing weights to apply to each xts within the list.}
}
\description{
Buckets factors into quantiles
}
