% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_rank.R
\name{factor_rank}
\alias{factor_rank}
\title{Nonparametric method for estimating factor scores.}
\usage{
factor_rank(
  factor_list,
  weights,
  format = "\%Y-\%m-\%d",
  ties.method = c("average", "first", "last", "random", "max", "min"),
  combine_function = "+"
)
}
\arguments{
\item{factor_list}{A list of equal length matrices or xts objects of factors}

\item{format}{The date format of the xts object, used to turn matrix back into #' xts}

\item{combine_function}{A string containing the name of the function used to combine weighted scores, if weights is specified. Defaults to sum scores.}

\item{weights:}{A list or vector with the same length as the number of columns #' in any xts or matrix object within factor_list. This list or #'  vector should sum #' to one. Alternatively, an xts object can be specified, in which case the #'dimensions of weights must match those of any #'  matrix or xts within factor_list. #'Each row of the weights must sum to one. The list or column index of weights #'corresponds to the index of #'  the factor_list to which the weights refer.}
}
\value{
Returns a list of ranked matrix or xts objects if weights is not #'specified.

If weights is specified, returns a single matrix or xts containing a #'weighted sum of factor scores.
}
\description{
Nonparametric method for estimating factor scores.
}
\references{
DiStefano, C., Zhu, M., & Mîndrilã, D. (2009). Understanding and Using Factor Scores:
Considerations for the Applied Researcher. \emph{Practical Assessment, Research, and Evaluation}, 14(20).
}
