% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorapply.R
\name{factorapply}
\alias{factorapply}
\title{Creates a list of matrices or time series resulting from a list of different functions.}
\usage{
factorapply(price, functions, arguments, lag = FALSE)
}
\arguments{
\item{price}{A numeric matrix or xts object.}

\item{arguments}{A list of lists containing corresponding arguments to each function. If no arguments are specified, an empty list must exist for that function.}

\item{lag}{Specifies whether or not to lag all columns by one. Defaults to true.}

\item{factors}{A list containing strings corresponding to function names to call.}
}
\value{
Returns a list of xts resulting from the functions specified in the factors list.
}
\description{
Creates a list of matrices or time series resulting from a list of different functions.
}
