% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_max_lookback.R
\name{trim_max_lookback}
\alias{trim_max_lookback}
\title{Trims the maximum number of leading NAs from every matrix in a list of matrices.}
\usage{
trim_max_lookback(factor_list)
}
\arguments{
\item{factor_list}{A list of matrices or xts objects of equal length.}
}
\value{
A list of matrices or xts objects of equal length without NAs.
}
\description{
Trims the maximum number of leading NAs from every matrix in a list of matrices.
}
\note{
This functions ensures that all factors contribute to the factor score in factor_scores regardless of lookback period.

All other functions outside of factorapply depend on this structure.
}
